Rename repository files and update imports
