+++ src/repositories/earningRepository.ts
